@@ -24,9 +24,9 @@ import static com.example.forum.util.QueryUtil.getMessageboardId;
 
 @Repository
 public class ForumDao {
-    private SimpleJdbcQuery simpleJdbcQuery;
-    private SimpleJdbcInsert messageJdbcInsert;
-    private SimpleJdbcInsert messageBoardJdbcInsert;
+    private final SimpleJdbcQuery simpleJdbcQuery;
+    private final SimpleJdbcInsert messageJdbcInsert;
+    private final SimpleJdbcInsert messageBoardJdbcInsert;
 
     public ForumDao(DataSource dataSource) {
         this.simpleJdbcQuery = new SimpleJdbcQuery(dataSource);
@@ -34,12 +34,22 @@ public class ForumDao {
         this.messageBoardJdbcInsert = new SimpleJdbcInsert(dataSource).withTableName("messageboard").usingGeneratedKeyColumns("messageboard_id");
     }
 
-    EnhancedRowMapper<Message> messageRowMapper = rs -> Message.builder()
+    /*
+    Note: With builder pattern the code would be less prone to errors with the order of constructor fields.
+          On the other hand, the builder is boilerplate.
+
+    private EnhancedRowMapper<Message> messageRowMapper = rs -> Message.builder()
             .messageId(getMessageId(rs))
             .userId(rs.getInt("user_id"))
             .createTime(rs.getTimestamp("create_time").toLocalDateTime())
             .text(rs.getString("text"))
             .build();
+     */
+    private final EnhancedRowMapper<Message> messageRowMapper = rs -> new Message(
+            getMessageId(rs),
+            rs.getInt("user_id"),
+            rs.getString("text"),
+            rs.getTimestamp("create_time").toLocalDateTime());
 
     public List<Message> getMessages(MessageboardId messageboardId) {
         return simpleJdbcQuery
@@ -58,11 +68,10 @@ public class ForumDao {
     public List<Messageboard> getMessageboards() {
         var counts = getMessageCounts();
         return simpleJdbcQuery.queryStr("select * from messageboard order by messageboard_id")
-                .listWith(rs -> Messageboard.builder()
-                                .messageboardId(getMessageboardId(rs))
-                                .name(rs.getString("name"))
-                                .messageCount(counts.getOrDefault(getMessageboardId(rs), 0L))
-                                .build());
+                .listWith(rs -> new Messageboard(
+                        getMessageboardId(rs),
+                        rs.getString("name"),
+                        counts.getOrDefault(getMessageboardId(rs), 0L)));
     }
 
     public Map<MessageboardId, Long> getMessageCounts() {
@@ -79,9 +88,9 @@ public class ForumDao {
 
     public Message insertMessage(MessageboardId messageboardId, InsertMessage insertMessage) {
         MapSqlParameterSource params = new MapSqlParameterSource()
-                .addValue("messageboard_id", messageboardId.intValue)
-                .addValue("user_id", insertMessage.userId)
-                .addValue("text", insertMessage.text)
+                .addValue("messageboard_id", messageboardId.intValue())
+                .addValue("user_id", insertMessage.userId())
+                .addValue("text", insertMessage.text())
                 .addValue("create_time", LocalDateTime.now());
 
         int id = messageJdbcInsert.executeAndReturnKey(params).intValue();
@@ -90,7 +99,7 @@ public class ForumDao {
 
     public MessageboardId insertMessageboard(InsertMessageboard insertMessageboard) {
         MapSqlParameterSource params = new MapSqlParameterSource()
-                .addValue("name", insertMessageboard.name);
+                .addValue("name", insertMessageboard.name());
 
         int id = messageBoardJdbcInsert.executeAndReturnKey(params).intValue();
         return new MessageboardId(id);
@@ -106,7 +115,7 @@ public class ForumDao {
     public int editMessage(MessageboardId messageboardId, MessageId messageId, EditMessage editMessage) {
         return simpleJdbcQuery
                 .queryStr("update message set text = ? where messageboard_id = ? and message_id = ?")
-                .params(editMessage.text, messageboardId, messageId)
+                .params(editMessage.text(), messageboardId, messageId)
                 .execute();
     }
 }
